refactor: 바텀시트와 모달 컴포넌트 위치 변경
