+++ src/components/itemstore/StoreItemList.tsx
@@ -1,84 +1,30 @@
 'use client';
 
-import useBottomSheetStore from '@/stores/useBottomSheetStore';
-import BottomSheet from '../ui/bottomsheet/BottomSheet';
-import ConfirmItemPurchase from './ConfirmItemPurchase';
-import { useState } from 'react';
 import { Item } from '@/types/Item.types';
 import { useStoreItems } from '@/hooks/queries/useStoreItems';
 import StoreItemCard from '@/components/itemstore/StoreItemCard';
-import { useUpdateMyItem } from '@/hooks/mutations/useUpdateMyItem';
-import { isPointSufficient } from '@/services/itemstore/isPointSufficient';
-import { isItemOwned } from '@/services/itemstore/isItemOwned';
-import useActionModalStore from '@/stores/useActionModalStore';
-import { useMediaQuery } from 'react-responsive';
-import ActionModal from '../ui/modal/ActionModal';
 
 type StoreItemListProps = {
-  point: number;
-  userId: string;
-  myItems: Item[];
+  onItemSelect: (item: Item) => void;
 };
 
-const StoreItemList = ({ point, userId, myItems }: StoreItemListProps) => {
-  const isDesktop = useMediaQuery({ minWidth: '1440px' });
-  const { open: openBottomSheet, close: closeBottomSheet } = useBottomSheetStore();
-  const { open: openActionModal, close: closeActionModal } = useActionModalStore();
+const StoreItemList = ({ onItemSelect }: StoreItemListProps) => {
   const { data: storeItems, isLoading } = useStoreItems();
-  const [selectedItem, setSelectedItem] = useState<Item | null>(null);
-  const { mutate: updateMyItem } = useUpdateMyItem(userId);
-
-  const onClickHandler = (item: Item) => {
-    updateMyItem(item);
-    if (isDesktop) {
-      closeActionModal();
-    } else {
-      closeBottomSheet();
-    }
-  };
 
   if (isLoading || !storeItems) return '로딩 중입니다...';
 
   return (
-    <>
-      <div className='flex gap-[13px] desktop:gap-[35px] overflow-x-auto whitespace-nowrap min-h-[165px] desktop:min-h-[429px] -mr-[20px] first:pl-[2px] last:pr-[20px]'>
-        {storeItems.map((item) => (
-          <div
-            key={item.id}
-            onClick={() => {
-              setSelectedItem(item);
-              if (isDesktop) {
-                openActionModal();
-              } else {
-                openBottomSheet();
-              }
-            }}
-            className='cursor-pointer'
-          >
-            <StoreItemCard {...item} />
-          </div>
-        ))}
-      </div>
-      {isDesktop ? (
-        <ActionModal
-          onClick={() => onClickHandler(selectedItem!)}
-          type='confirm'
-          label='아이템 구매하기'
-          disabled={!selectedItem || !isPointSufficient(selectedItem, point) || isItemOwned(selectedItem, myItems)}
+    <div className='flex gap-[13px] desktop:gap-[35px] overflow-x-auto whitespace-nowrap min-h-[165px] desktop:min-h-[429px] -mr-[20px] first:pl-[2px] last:pr-[20px]'>
+      {storeItems.map((item) => (
+        <div
+          key={item.id}
+          onClick={() => onItemSelect(item)}
+          className='cursor-pointer'
         >
-          {selectedItem && <ConfirmItemPurchase selectedItem={selectedItem} />}
-        </ActionModal>
-      ) : (
-        <BottomSheet
-          onClick={() => onClickHandler(selectedItem!)}
-          type='confirm'
-          label='아이템 구매하기'
-          disabled={!selectedItem || !isPointSufficient(selectedItem, point) || isItemOwned(selectedItem, myItems)}
-        >
-          {selectedItem && <ConfirmItemPurchase selectedItem={selectedItem} />}
-        </BottomSheet>
-      )}
-    </>
+          <StoreItemCard {...item} />
+        </div>
+      ))}
+    </div>
   );
 };
